+++ packages/framework/src/Modules/Metadata/MetadataBag.php
@@ -5,7 +5,7 @@ namespace Hyde\Framework\Modules\Metadata;
 use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Helpers\Meta;
 use Hyde\Framework\Models\Pages\MarkdownPost;
-use Hyde\Framework\Modules\Metadata\Models\BaseMetadataElement;
+use Hyde\Framework\Modules\Metadata\Models\GenericMetadataElement;
 use Illuminate\Contracts\Support\Htmlable;
 
 /**
@@ -50,7 +50,7 @@ class MetadataBag implements Htmlable
         );
     }
 
-    public function add(BaseMetadataElement|string $item): static
+    public function add(GenericMetadataElement|string $item): static
     {
         if ($item instanceof Models\LinkElement) {
             $this->links[$item->uniqueKey()] = $item;