+++ packages/framework/src/Modules/Metadata/Models/GenericMetadataElement.php
@@ -2,7 +2,7 @@
 
 namespace Hyde\Framework\Modules\Metadata\Models;
 
-abstract class BaseMetadataElement implements \Stringable
+abstract class GenericMetadataElement implements \Stringable
 {
     abstract public function __toString(): string;